Organizes login form and schema into dedicated folders
